Rewrite Progress with TypeScript
