+++ src/elements/Progress/index.ts
@@ -1,0 +1,1 @@
+export { default as Progress } from './Progress';
